internal/postgres: change version_map.module_id insert
We're still seeing issues with UpsertVersionMap, even though the queue
has been dialed down to insert only one module at a time.

Change UpsertVersionMap to be two queries:
(1) select moduleID: should be a relatively trivial query that just
fetches the module_id for a given path and version from the modules
table
(2) upsert version_map: same as query before the data model changes,
which upserts a row in the version_map table, but now it also upserts
a module_id

This will allow us to get more information on where in UpsertVersionMap
the query is failing.

Not that both the current query on master and the new one succeed when
running the worker locally and connecting to the dev database.

Change-Id: I3c1d0a621294fa0e38bdd2165a35460dbacda4e6
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/719980
Reviewed-by: Jonathan Amsterdam <jba@google.com>