internal/postgres: do not delete package version states
At the moment when we upsert a row into module_version_states, we
delete and repopulate existing rows in package_version_states.

This adds an additional query that is costly and isn't necessary, since
we wouldn't expect there to be packages missing in a future round of
processing. Also, if for some reason that did happen, it would be useful
to have this data captured in package_version_states anyways.

Change-Id: Iceb64235957ade115293bb7c1ff62cdcb2e2073f
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/725870
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>