internal/frontend: fix a server test
Before the recent changes to sample, it was easy for the V1Path of a
package to be out of sync with the package path. This happened in
the setup for TestServer, causing the version page to include a
"Versions in this module" header that shouldn't have been there. (The
version logic thought there was another module with the package,
because of the bad V1Ppath.)

Also, output the HTML on error for easier debugging.

Change-Id: Ibf29e3c90dc1c12f372946d7e6615f7b610dc838
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/733606
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>