internal/postgres: add requeue limit for large modules
Our requeue logic is such that large modules (ones with > 1500 packages)
are requeued last, since they can take a long time to reprocess.
However, with large modules, it's possible that the modules that are
queued do not finish processing before the dedupe period for Cloud Tasks
is up, causing the queue to be backed-up with duplicate modules.

A limit of is now added so that at most 100 large modules can be added
to the queue at once.

Change-Id: I0d2a9b709b9a9972b609e9c3af125d58d37f4488
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/755584
Reviewed-by: Jonathan Amsterdam <jba@google.com>