internal/fetch/dochtml: render shortened func signatures in side nav
This change introduces a new method on Renderer, ShortSynopsis. It
is meant to be used in templates when rendering methods and functions
within the side navigation UI component.

Due to the limited horizontal space available, it omits the “func”
keyword, type information, and return values.

An example of functions and methods from the time package rendered
using this new method is below.

Functions
    After(d)
    Sleep(d)
    Tick(d)
Types
    type Duration
        ParseDuration(s)
        Since(t)
        Until(t)
        (d) Hours()
        (d) Microseconds()
        (d) Milliseconds()
        (d) Minutes()
        (d) Nanoseconds()
        (d) Round(m)
        (d) Seconds()
        (d) String()
        (d) Truncate(m)
    type Location
        FixedZone(name, offset)
        LoadLocation(name)
        LoadLocationFromTZData(name, data)
        (l) String()
    type Month
        (m) String()
    type ParseError
        (e) Error()

Updates b/148095016

Change-Id: I663eaafdc0baa3619e449ccd0d8ee8d601974392
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/763562
CI-Result: Cloud Build <devtools-proctor-result-processor@system.gserviceaccount.com>
Reviewed-by: Julie Qiu <julieqiu@google.com>