internal/frontend: enqueue fetch request for all requests to path@master
Once we turn on the feature flag for supporting requests to master, we
need a way to ensure that master is not stale for a given path.

As a result, all requests to path@master will be enqueued, once we
determine that it is a valid path.

Updates golang/go#36811

Change-Id: I498179586ee46acbd7347eb70a90104fe9e0b8a8
Reviewed-on: https://team-review.git.corp.google.com/c/golang/discovery/+/769967
Reviewed-by: Jonathan Amsterdam <jba@google.com>