internal/teeproxy: add metrics for requests
Add metrics for request latencies and counts from godoc.org and to
pkg.go.dev, as well as total server metrics.

A custom HTTP status code is added to denote that the circuit breaker
rejected the request. This status code is necessary because the metrics
are aggregated by status code, and we need to differentiate a rate
limited request vs a request rejected by the circuit breaker.

Fixes golang/go#39843

Change-Id: I16aa34cf59794e1d226b5a77fbafca34da47ff81
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/239818
Reviewed-by: Julie Qiu <julie@golang.org>