internal/frontend: trust bluemonday
Rewrite bluemonday's HTML sanitization in a way that makes it
clear that we are trusting bluemonday.

There's no functional change here, just a slight reorg and a
replacement of a "Riskily..." function with a "KnownTo..." one to mark
that we have considered the sanitization we're applying and trust it
to be safe.

Change-Id: Ibbcb64428f76875f8c95734633ca6efff8e387f8
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/240502
Reviewed-by: Julie Qiu <julie@golang.org>