internal/frontend: return 400 on @master requests for stdlib paths
At the moment, we do not support requests to @master for packages in the
standard library. Until that support is added, return a 400 for those
requests, instead of showing a 404 page where the user could attempt to
fetch the path.

fetchAndPoll is also updated to return a 400 on requests to @master and
@latest for packages in the standard library.

Updates golang/go#36811
Updates golang/go#39973

Change-Id: I372ad2686547278754f2c04e69ddea7fc092df34
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/240682
Reviewed-by: Jonathan Amsterdam <jba@google.com>