content/static: select side navigation element when scrolling
This change ensures that the topmost element in view has its
corresponding navigation element selected. This is done via
two IntersectionObservers: one to handle the target element and
one to handle all the rest. This is due to the increased size of
the target element to ensure it displays properly under the
sticky header.

Various optimizations are in place to ensure that the
IntsersectionObserver callback does minimal querying of the DOM.
An up front cost of an initial indexing operation in the navigation
controller is a candidate for optimization in future work.

Change-Id: I339085e03efb4708e2739488a36cfab9cd8f1821
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/240689
Reviewed-by: Alexander Rakoczy <alex@golang.org>