internal/frontend: specify requested or resolved version in args
Many functions in internal/frontend have "version" as an argument, but
it is not clear whether that version is the "requested version" (i.e.
latest, master, or semver), or the "resolved version" (the semantic
version corresponding to the requested version).

These are now specified in the argument name to make it clearer what is
expected as an input to the function.

Change-Id: I9211c1b55e0951b37f693bfed5c4b509a5d81da4
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/240939
Reviewed-by: Jonathan Amsterdam <jba@google.com>