doc: improve 'go run' tips for cmd packages
Running a single Go file is OK if that file is specifically standalone,
such as gen_foo.go. However, if we're building an entire package, it's
both easier and better to build the entire package instead.

Even if the package is currently just one file, using 'go run
./cmd/frontend' is easier to type, and won't break if the package ever
gets more Go files, which is likely.

While at it, use indentation for short single-line code blocks, since
that's nicer in plaintext than triple backticks.

Change-Id: If905cbe79eedbd676dbeb4ead2588a9a6a7ebb76
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/241018
Reviewed-by: Julie Qiu <julie@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>