internal/frontend: fix serveDetails for master version and latest error
LegacyGetPackage and LegacyGetModuleInfo do not support fetching the
master version of a package. Rather than changing the queries in those
functions, we can use the resolvedVersion from the GetPathInfo in
serveDetails to get that data. This is now passed in as an argument to
servePackagePage and serveModulePage.

Additionally, logic for serving the various NotFound pages is moved into
a single servePathNotFoundPage function, which is used when
internal.ExperimentUsePathInfo is on.

Updates golang/go#36811

Change-Id: Ic4f3f6bbf211affa98891ca3a25c8219913e38ec
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/240945
Reviewed-by: Jonathan Amsterdam <jba@google.com>