internal/fetch/dochtml: render doc safely
Render doc for declarations with safehtml.

We do this the same way we handled example HTML: we render
text into a constant template.

Keeping the template formatted nicely required some changes in the
result, but they only affect irrelevant whitespace.

We have to convert to template.HTML at the last moment before
inserting into the template, or we'll double-escape. That
will go away when we finally switch to use safehtml/template.

Change-Id: I2ebd753d3c8692be1e829079d02fad89eddf10e7
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/241937
Reviewed-by: Julie Qiu <julie@golang.org>