internal/frontend: add serveDetailsPage
serveDetailsPage is added, which replaces servePackagePageNew in serving
details pages when the "use-directories" experiment flag is on. Like
servePackagePageNew, serveDetailsPages supports the package/directory
views. It also supports the module view, so that we can stop using
LegacyModuleInfo and reading from modules.readme_file_path and
modules.readme_contents.

For golang/go#39629

Change-Id: I8e664bf1e9174a630db6b723c949ba1e9aa0dc9b
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/241397
Reviewed-by: Jonathan Amsterdam <jba@google.com>