internal/fetch: add versioning for import packages from same module
The HTML documentation for packages currently links imported types and
packages to the latest version of the imported package. Link these to
the same version as the current module instead if the imported package
belongs to the same module.

In the future, we may want to use the go.mod file or the go list command
to link all imports (and not just packages in the same module) to the
same version that the author intended.

Fixes golang/go#38429

Change-Id: If8cc412e11c626003c16595a7037a6c80e728b02
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/241818
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>