internal/frontend: update legacy and new terminology
The use of "legacy" and "new" is updated in the frontend package, so
that "legacy" is used as prefix to any function that will be removed once
the data migration is completed, or currently contains legacy types.
Otherwise, the function will remain after the data migration. The "new"
suffix is removed from these functions if it was present.

For golang/go#39629

Change-Id: Iae88fef8c53073442a8eeb0eeb54a3296448949e
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/241839
Reviewed-by: Jonathan Amsterdam <jba@google.com>