internal/frontend: fix logic for when to fetch path@master
An error is fixed in determining when to return a BadRequest for
/fetch/path@master. Previously, we were returning a 400 if the
"master-version" flag is on, and the requested version is not "master.

The logic should be such that we return a 400 if the requested version
is master, unless the "master-version" flag is on.

For golang/go#36811
For golang/go#37002

Change-Id: Id080dc27e11e144dfdc903f884a8978cebf4c15c
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/242358
Reviewed-by: Jamal Carvalho <jamal@golang.org>