internal/middleware: update quota server to check AuthHeader and AcceptedAuthValues
The quota server now checks for authentication using a known
header, to see if requests from the source should be rate-limited.

This replaces an existing implementation, where the quota server checked
for the "Referer", since this is header is usually not set by requests
that do not originate from a browser.

This functionality is used to allow the teeproxy to forward requests to
pkg.go.dev.

Change-Id: Ic1eac95c3047219842b561f2cee3b0547fe36c5c
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/243283
Reviewed-by: Jonathan Amsterdam <jba@google.com>