internal/frontend: use escaped path when modifying image sources
When walking HTML inside markdown to translate relative image sources,
use the escaped path of the source rather than the original path.

Previously, if the value of an img src tag was escaped, like

    a%20b

it was being converted to

    a b

The sanitizer then removed the entire img tag.

Using URL.EscapedPath preserves the original escaping.

Change-Id: I05544a7f2d089c7026dedac605e07b193823957b
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/243618
Reviewed-by: Julie Qiu <julie@golang.org>