internal/stdlib: fix stdlib README for new data model
The overview tab for the stdlib package is currently displaying
src/README.vendor, instead of README.md, when the use-directories
experiment flag is on. This is because based on the way the std
module zip is constructed, both README.md and README.vendor are present
at the root for version v1.4.0-beta.1 and higher. The
fetch process will then choose README.vendor to include, since
it comes after README.md alphabetically.

The README.vendor file is no longer included in the std zip.

For golang/go#39629

Change-Id: I1c9d8fcd1776f7b58be08f5f1ec1dc1b735e736d
Reviewed-on: https://go-review.googlesource.com/c/pkgsite/+/244608
Reviewed-by: Jonathan Amsterdam <jba@google.com>